[ADD] New GUI controls
This commit introduces new GUI controls using ImGui for SFML backend,
allowing the user to modify the simulation parameters like gravity,
damp or the pendulum parameters like the rod length and bob mass.

Additionally, this commit is aimed at cleaning up a lot of stuff and
getting a good final design of the project and modularising most of the
components.