[ADD] State select slider to choose current state
The state select slider can be used to select the current state
which can be modified. Clicking on reset state will set the state
of the current selected state to the values in the fields.

This can be used to change restart a particular state with a different
mass or rod length or even initial angle.